% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_match.R
\name{geo_match}
\alias{geo_match}
\title{Match case locations to pollution grid}
\usage{
geo_match(cases, pollution)
}
\arguments{
\item{cases}{Dataframe with columns: id, date, lon, lat}

\item{pollution}{Pollution data from load_pollution()}
}
\value{
Dataframe with added exposure column
}
\description{
Match case locations to pollution grid
}
\examples{
matched_data <- geo_match(sample_cases, pm_data)
}
