% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loader.R
\name{load_pollution}
\alias{load_pollution}
\title{Load air pollution data from NetCDF file}
\usage{
load_pollution(nc_path)
}
\arguments{
\item{nc_path}{Path to the NetCDF file}
}
\value{
A list containing lon, lat, time, and PM2.5 data
}
\description{
Load air pollution data from NetCDF file
}
\examples{
pm_data <- load_pollution("data/CHAP_PM2.5_20200101.nc")
}
